// src/features/admin/goods/components/GoodsVerificationForm/components/types.ts
export interface FormErrors {
  description?: string;
  length?: string;
  width?: string;
  height?: string;
  cbm?: string;
  weight?: string;
  quantity?: string;
  unitPrice?: string;
  location?: string;
  receivedByName?: string;
}

export interface FormData {
  description: string;
  length: string;
  width: string;
  height: string;
  cbm: string;
  weight: string;
  quantity: string;
  unitPrice: string;
  location: string;
  receivedByName: string;
}